Add `accept_array` option
This commit adds a new `accept_array` option for validating urls as
either a single string, or as an array of strings that may or may not
contain urls.

All other options continue to work as normal against any string in the
array.

This is useful for an object that takes advantage of Postgres's built
in Array type to hold URLs, which ActiveRecord supports natively.

```
validates :homepage, url: { accept_array: true }

validates :homepage, url: { accept_array: true, allow_blank: true}
```