Alway use given message ...
fixes #122.

I could only run the specs by removing Gemfile.lock and would recommend #150 to be merged before this PR.

Several examples were red with NoMethodErrors before I made any changes. I had to add the `value && ` to fix that.

The example `does not allow a url with only a scheme` was also red. I had to add the `host.empty?` check to fix that one.